Flattening libsuffixtools so that -I compiler options are no longer needed. The other way to do this would have been to find all the references to files in subdirectories and fix them up to actually use the subdirectories.

+++ libsuffixtools/MkqsThread.h
@@ -61,7 +61,7 @@ class MkqsThread
         sem_t* m_pQueueSem; // shared
         sem_t* m_pDoneSem; // shared
         
-        int m_thresholdSize;
+        size_t m_thresholdSize;
         const PrimarySorter* m_pPrimary;
         const FinalSorter* m_pFinal;
